add more tests for polygons, split pybind macros to specific bindings files
