+++ src/rasputin/cpp/raster_bindings.cpp
@@ -1,9 +1,13 @@
 #include <pybind11/pybind11.h>
+#include <pybind11/stl.h>
 #include <pybind11/numpy.h>
 
 #include "raster_data.h"
 
 namespace py = pybind11;
+
+PYBIND11_MAKE_OPAQUE(std::vector<rasputin::RasterData<float>>);
+PYBIND11_MAKE_OPAQUE(std::vector<rasputin::RasterData<double>>);
 
 template<typename T>
 void bind_raster_list(py::module &m, const std::string& pyname) {
